feat: 实现通用分页查询与DTO基类优化
refactor: 重构服务层与仓储层接口命名

fix: 修复TokenUtils中accessToken过期时间错误

feat(Query): 添加QueryExpressionBuilder智能查询构建器

feat(DTO): 新增BasePageDTO作为分页基类

feat(Controller): 实现通用CRUD控制器模板

feat(Json): 添加DatetimeJsonConverter统一日期格式

fix(Filter): 优化全局异常过滤器日志记录

feat(Autofac): 注册泛型仓储与服务

style: 清理无用using与注释

chore: 更新项目依赖与启动端口配置

+++ FastStart.Domain/Entity/SysRole.cs
@@ -44,14 +44,14 @@ namespace FastStart.Domain.Entity
         /// 默认值: 1
         ///</summary>
         [SugarColumn(ColumnName = "menu_check_strictly")]
-        public byte? MenuCheckStrictly { get; set; }
+        public int? MenuCheckStrictly { get; set; }
 
         /// <summary>
         /// 部门树选择项是否关联显示
         /// 默认值: 1
         ///</summary>
         [SugarColumn(ColumnName = "dept_check_strictly")]
-        public byte? DeptCheckStrictly { get; set; }
+        public int? DeptCheckStrictly { get; set; }
 
         /// <summary>
         /// 角色状态（0正常 1停用）
